pyproject.toml
[tool.poetry]
name = "jupiter-faq-bot"
version = "0.1.0"
description = "A conversational AI bot for Jupiter banking FAQs"
authors = ["Jupiter FAQ Bot <bot@jupiter.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
selenium = "^4.15.0"
webdriver-manager = "^4.0.0"
openai = "^1.0.0"
faiss-cpu = "^1.7.0"
chromadb = "^0.4.0"
streamlit = "^1.28.0"
pandas = "^2.1.0"
numpy = "^1.24.0"
python-dotenv = "^1.0.0"
jupyter = "^1.0.0"
scikit-learn = "^1.3.0"
nltk = "^3.8.0"
sentence-transformers = "^2.2.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"

[tool.poetry.scripts]
start = "streamlit run demo/streamlit_app.py --server.port 8000 --server.address 0.0.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
